worked on changes in restaurant list and restraunt details and added a new custom hook
