+++ src/Util/useRestrauntList.js
@@ -1,0 +1,20 @@
+import { useEffect, useState } from 'react';
+import { swiggyTopRestrauntAPI } from './Constant';
+const useRestrauntList = () =>{
+    const [listRestaurant, setListRestaurant] = useState();
+
+    useEffect(()=>{
+        fetchData()
+    }, [])
+    async function fetchData(){
+
+        const url = 'https://thingproxy.freeboard.io/fetch/' + encodeURIComponent(swiggyTopRestrauntAPI);
+        const restaurants = await fetch(url) ;
+        const restaurantJson = await restaurants.json();
+        const topRestaurant = restaurantJson?.data?.cards[1]?.card?.card?.gridElements?.infoWithStyle?.restaurants;
+        console.log(restaurantJson)
+        setListRestaurant(topRestaurant);
+    }
+    return (listRestaurant)
+}
+export default useRestrauntList;
